requirements.txt
SQLAlchemy==2.0.38
passlib==1.7.4
fastapi==0.115.10
uvicorn==0.34.0
psycopg2==2.9.10
python-multipart==0.0.20
PyJWT==2.10.1
bcrypt==4.3.0